Extracted class Waiter from iomediator00 into a separate module.
